move item group to main class
