+++ src/main/java/net/creatorsonion/paintingmod/entity/PMEntityRegistry.java
@@ -1,7 +1,6 @@
 package net.creatorsonion.paintingmod.entity;
 
 import net.creatorsonion.paintingmod.PaintingMod;
-import net.creatorsonion.paintingmod.entity.decoration.painting.DesignersPaintingEntity;
 import net.fabricmc.fabric.api.object.builder.v1.entity.FabricEntityTypeBuilder;
 import net.minecraft.entity.EntityDimensions;
 import net.minecraft.entity.EntityType;
